Change command request inheritance structure. Fix data races during requests event processing.
