+++ src/components/application_manager/include/application_manager/commands/command.h
@@ -115,7 +115,14 @@ class Command {
    * has exceed it's limit
    *
    */
-  virtual void onTimeOut() = 0;
+  virtual void HandleTimeOut() = 0;
+
+  /**
+   * @brief Function is called by RequestInfo when request controller
+   * updates request timeout
+   * Function sets request state to "AwaitingResponse"
+   */
+  virtual void OnUpdateTimeOut() = 0;
 
   /**
    * @brief AllowedToTerminate tells if request controller is allowed
