+++ src/components/application_manager/include/application_manager/commands/request_to_hmi.h
@@ -35,6 +35,7 @@
 
 #include "application_manager/application_manager.h"
 #include "application_manager/commands/command_impl.h"
+#include "application_manager/commands/command_request_impl.h"
 
 namespace application_manager {
 
@@ -61,7 +62,7 @@ bool ChangeInterfaceState(ApplicationManager& application_manager,
                           const smart_objects::SmartObject& response_from_hmi,
                           HmiInterfaces::InterfaceID interface);
 
-class RequestToHMI : public CommandImpl {
+class RequestToHMI : public CommandRequestImpl {
  public:
   RequestToHMI(const MessageSharedPtr& message,
                ApplicationManager& application_manager,
