+++ src/components/application_manager/include/application_manager/event_engine/event_dispatcher.h
@@ -64,6 +64,15 @@ class EventDispatcher {
   /*
    * @brief Unsubscribes the observer from specific event
    *
+   * @param event_id    The event ID to subscribe for
+   * @param hmi_correlation_id  The event HMI correlation ID
+   */
+  virtual void remove_observer(const Event::EventID& event_id,
+                               const int32_t hmi_correlation_id) = 0;
+
+  /*
+   * @brief Unsubscribes the observer from specific event
+   *
    * @param event_id    The event ID to unsubscribe from
    * @param observer    The observer to be unsubscribed
    */
