+++ src/components/application_manager/include/application_manager/event_engine/event_dispatcher_impl.h
@@ -83,6 +83,9 @@ class EventDispatcherImpl : public EventDispatcher {
    */
   void raise_event(const Event& event) OVERRIDE;
 
+  void remove_observer(const Event::EventID& event_id,
+                       const int32_t hmi_correlation_id) OVERRIDE;
+
   /*
    * @brief Subscribe the observer to event
    *
