+++ src/components/application_manager/include/application_manager/request_controller.h
@@ -47,6 +47,7 @@
 #include "interfaces/HMI_API.h"
 #include "interfaces/MOBILE_API.h"
 
+#include "application_manager/event_engine/event_dispatcher_impl.h"
 #include "application_manager/request_controller_settings.h"
 #include "application_manager/request_info.h"
 #include "application_manager/request_tracker.h"
@@ -87,7 +88,8 @@ class RequestController {
    * @brief Class constructor
    *
    */
-  RequestController(const RequestControlerSettings& settings);
+  RequestController(const RequestControlerSettings& settings,
+                    event_engine::EventDispatcher& event_disptacher);
 
   /**
    * @brief Class destructor
@@ -134,6 +136,36 @@ class RequestController {
    * @param ptr Reference to shared pointer that point on hmi notification
    */
   void addNotification(const RequestPtr ptr);
+
+  /**
+   * @brief RetainRequestInstance retains request instance by its
+   * connection+correlation key
+   * @param connection_key connection key of application
+   * @param correlation_id correlation id of request
+   * @return true if request was rerained. false if the request with such
+   * connection+correlation key was not found
+   */
+  bool RetainRequestInstance(const uint32_t connection_key,
+                             const uint32_t correlation_id);
+
+  /**
+   * @brief RemoveRetainedRequest removes request instance retained before
+   * @param connection_key connection key of application
+   * @param correlation_id correlation id of request
+   */
+  void RemoveRetainedRequest(const uint32_t connection_key,
+                             const uint32_t correlation_id);
+
+  /**
+   * @brief IsStillWaitingForResponse check if request is still waiting for
+   * response
+   * @param connection_key connection key of application
+   * @param correlation_id correlation id of request
+   * @return true if request is still waiting for response, otherwise returns
+   * false
+   */
+  bool IsStillWaitingForResponse(const uint32_t connection_key,
+                                 const uint32_t correlation_id) const;
 
   /**
    * @brief Removes request from queue
@@ -271,12 +303,18 @@ class RequestController {
   std::list<RequestPtr> mobile_request_list_;
   sync_primitives::Lock mobile_request_list_lock_;
 
-  /*
-   * Requests, that are waiting for responses
+  /**
+   * @brief Requests, that are waiting for responses
    * RequestInfoSet provides correct processing of requests with thre same
    * app_id and corr_id
    */
   RequestInfoSet waiting_for_response_;
+
+  /**
+   * @brief Requests, that are retained to be not destroyed right after
+   * sending response to mobile request
+   */
+  HashSortedRequestInfoSet retained_mobile_requests_;
 
   /**
    * @brief Tracker verifying time scale and maximum requests amount in
@@ -310,6 +348,7 @@ class RequestController {
 
   bool is_low_voltage_;
   const RequestControlerSettings& settings_;
+  event_engine::EventDispatcher& event_dispatcher_;
   DISALLOW_COPY_AND_ASSIGN(RequestController);
 };
