+++ src/components/application_manager/rpc_plugins/app_service_rpc_plugin/include/app_service_rpc_plugin/commands/hmi/as_get_app_service_data_request_from_hmi.h
@@ -71,7 +71,7 @@ class ASGetAppServiceDataRequestFromHMI
   /**
    * @brief onTimeOut from request controller
    */
-  virtual void onTimeOut();
+  virtual void OnTimeOut();
 
   /**
    * @brief on_event allows to handle events
