+++ src/components/application_manager/src/event_engine/event_observer.cc
@@ -75,7 +75,7 @@ void EventObserver::unsubscribe_from_all_mobile_events() {
   event_dispatcher_.remove_mobile_observer(*this);
 }
 
-void EventObserver::on_event(const event_engine::MobileEvent& event) {}
+void EventObserver::HandleOnEvent(const event_engine::MobileEvent& event) {}
 
 }  // namespace event_engine
 }  // namespace application_manager
