+++ src/components/application_manager/src/request_controller.cc
@@ -30,12 +30,12 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include "utils/logger.h"
-
-#include "application_manager/commands/command_request_impl.h"
-#include "application_manager/commands/request_to_hmi.h"
 #include "application_manager/request_controller.h"
 
+#include "application_manager/commands/command_request_impl.h"
+#include "application_manager/commands/request_from_mobile_impl.h"
+#include "application_manager/commands/request_to_hmi.h"
+#include "utils/logger.h"
 #include "utils/timer_task_impl.h"
 
 namespace application_manager {
@@ -46,7 +46,9 @@ using namespace sync_primitives;
 
 SDL_CREATE_LOG_VARIABLE("RequestController")
 
-RequestController::RequestController(const RequestControlerSettings& settings)
+RequestController::RequestController(
+    const RequestControlerSettings& settings,
+    event_engine::EventDispatcher& event_dispatcher)
     : pool_state_(UNDEFINED)
     , pool_size_(settings.thread_pool_size())
     , request_tracker_(settings)
@@ -56,7 +58,8 @@ RequestController::RequestController(const RequestControlerSettings& settings)
                  this, &RequestController::TimeoutThread))
     , timer_stop_flag_(false)
     , is_low_voltage_(false)
-    , settings_(settings) {
+    , settings_(settings)
+    , event_dispatcher_(event_dispatcher) {
   SDL_LOG_AUTO_TRACE();
   InitializeThreadpool();
   timer_.Start(0, timer::kSingleShot);
@@ -233,6 +236,45 @@ void RequestController::removeNotification(
   SDL_LOG_DEBUG("Cannot find notification");
 }
 
+bool RequestController::RetainRequestInstance(const uint32_t connection_key,
+                                              const uint32_t correlation_id) {
+  SDL_LOG_AUTO_TRACE();
+  auto request = waiting_for_response_.Find(connection_key, correlation_id);
+  if (request) {
+    retained_mobile_requests_.insert(request);
+    SDL_LOG_DEBUG("Request (" << connection_key << ", " << correlation_id
+                              << ") has been retained");
+  }
+  SDL_LOG_DEBUG(
+      "Total retained requests: " << retained_mobile_requests_.size());
+
+  return static_cast<bool>(request);
+}
+
+void RequestController::RemoveRetainedRequest(const uint32_t connection_key,
+                                              const uint32_t correlation_id) {
+  SDL_LOG_AUTO_TRACE();
+  for (auto it = retained_mobile_requests_.begin();
+       it != retained_mobile_requests_.end();
+       ++it) {
+    if ((*it)->request()->connection_key() == connection_key &&
+        (*it)->request()->correlation_id() == correlation_id) {
+      SDL_LOG_DEBUG("Removing request (" << connection_key << ", "
+                                         << correlation_id << ")");
+      retained_mobile_requests_.erase(it);
+      break;
+    }
+  }
+  SDL_LOG_DEBUG(
+      "Total retained requests: " << retained_mobile_requests_.size());
+}
+
+bool RequestController::IsStillWaitingForResponse(
+    const uint32_t connection_key, const uint32_t correlation_id) const {
+  auto request = waiting_for_response_.Find(connection_key, correlation_id);
+  return static_cast<bool>(request);
+}
+
 void RequestController::TerminateRequest(const uint32_t correlation_id,
                                          const uint32_t connection_key,
                                          const int32_t function_id,
@@ -268,9 +310,13 @@ void RequestController::TerminateRequest(const uint32_t correlation_id,
     return;
   }
   if (force_terminate || request->request()->AllowedToTerminate()) {
+    event_dispatcher_.remove_observer(
+        static_cast<hmi_apis::FunctionID::eType>(function_id), correlation_id);
     waiting_for_response_.RemoveRequest(request);
   } else {
-    SDL_LOG_WARN("Request was not terminated");
+    SDL_LOG_WARN("Request was not terminated "
+                 << "correlation_id = " << correlation_id
+                 << " function_id = " << function_id);
   }
   NotifyTimer();
 }
@@ -309,9 +355,27 @@ void RequestController::terminateWaitingForExecutionAppRequests(
 void RequestController::terminateWaitingForResponseAppRequests(
     const uint32_t& app_id) {
   SDL_LOG_AUTO_TRACE();
-  waiting_for_response_.RemoveByConnectionKey(app_id);
-  SDL_LOG_DEBUG(
-      "Waiting for response count : " << waiting_for_response_.Size());
+  SDL_LOG_DEBUG("Waiting for response count before cleanup: "
+                << waiting_for_response_.Size());
+
+  auto pending_requests =
+      waiting_for_response_.GetRequestsByConnectionKey(app_id);
+  for (auto it = pending_requests.begin(); it != pending_requests.end(); ++it) {
+    auto request_ptr = (*it)->request();
+    if (!request_ptr->CleanUp()) {
+      SDL_LOG_WARN("Request with corr_id: "
+                   << request_ptr->correlation_id()
+                   << " can't be terminated right now. Keep in the queue");
+      continue;
+    }
+
+    SDL_LOG_DEBUG(
+        "Terminating request with corr_id: " << request_ptr->correlation_id());
+    waiting_for_response_.RemoveRequest(*it);
+  }
+
+  SDL_LOG_DEBUG("Waiting for response count after cleanup: "
+                << waiting_for_response_.Size());
 }
 
 void RequestController::terminateAppRequests(const uint32_t& app_id) {
@@ -353,7 +417,6 @@ void RequestController::updateRequestTimeout(const uint32_t& app_id,
   SDL_LOG_DEBUG(
       "New_timeout is NULL. RequestCtrl will "
       "not manage this request any more");
-
   RequestInfoPtr request_info =
       waiting_for_response_.Find(app_id, correlation_id);
   if (request_info) {
@@ -427,18 +490,23 @@ void RequestController::TimeoutThread() {
                  << " request id: " << probably_expired->requestId()
                  << " connection_key: " << probably_expired->app_id()
                  << " is expired");
-    const uint32_t experied_request_id = probably_expired->requestId();
-    const uint32_t experied_app_id = probably_expired->app_id();
+    const uint32_t expired_request_id = probably_expired->requestId();
+    const uint32_t expired_app_id = probably_expired->app_id();
 
-    probably_expired->request()->onTimeOut();
+    probably_expired->request()->HandleTimeOut();
+
     if (RequestInfo::HmiConnectionKey == probably_expired->app_id()) {
+      const uint32_t function_id = probably_expired->request()->function_id();
+      event_dispatcher_.remove_observer(
+          static_cast<hmi_apis::FunctionID::eType>(function_id),
+          expired_request_id);
       SDL_LOG_DEBUG("Erase HMI request: " << probably_expired->requestId());
       waiting_for_response_.RemoveRequest(probably_expired);
     }
     probably_expired = waiting_for_response_.FrontWithNotNullTimeout();
     if (probably_expired) {
-      if (experied_request_id == probably_expired->requestId() &&
-          experied_app_id == probably_expired->app_id()) {
+      if (expired_request_id == probably_expired->requestId() &&
+          expired_app_id == probably_expired->app_id()) {
         SDL_LOG_DEBUG("Expired request wasn't removed");
         break;
       }
@@ -490,8 +558,8 @@ void RequestController::Worker::threadMain() {
         std::make_shared<MobileRequestInfo>(request_ptr, timeout_in_mseconds);
 
     if (!request_controller_->waiting_for_response_.Add(request_info_ptr)) {
-      commands::CommandRequestImpl* cmd_request =
-          dynamic_cast<commands::CommandRequestImpl*>(request_ptr.get());
+      commands::RequestFromMobileImpl* cmd_request =
+          dynamic_cast<commands::RequestFromMobileImpl*>(request_ptr.get());
       if (cmd_request != NULL) {
         uint32_t corr_id = cmd_request->correlation_id();
         request_controller_->duplicate_message_count_lock_.Acquire();
