+++ src/components/application_manager/src/request_info.cc
@@ -92,6 +92,7 @@ void application_manager::request_controller::RequestInfo::updateEndTime() {
 void RequestInfo::updateTimeOut(const uint64_t& timeout_msec) {
   timeout_msec_ = timeout_msec;
   updateEndTime();
+  request_->OnUpdateTimeOut();
 }
 
 bool RequestInfo::isExpired() {
@@ -153,7 +154,7 @@ bool RequestInfoSet::Add(RequestInfoPtr request_info) {
 }
 
 RequestInfoPtr RequestInfoSet::Find(const uint32_t connection_key,
-                                    const uint32_t correlation_id) {
+                                    const uint32_t correlation_id) const {
   RequestInfoPtr result;
 
   // Request info for searching in request info set by log_n time
@@ -195,6 +196,22 @@ RequestInfoPtr RequestInfoSet::FrontWithNotNullTimeout() {
     }
   }
   return result;
+}
+
+std::list<RequestInfoPtr> RequestInfoSet::GetRequestsByConnectionKey(
+    const uint32_t connection_key) {
+  SDL_LOG_AUTO_TRACE();
+  sync_primitives::AutoLock lock(pending_requests_lock_);
+
+  std::list<RequestInfoPtr> output_list;
+  AppIdCompararator comparator(AppIdCompararator::Equal, connection_key);
+
+  std::copy_if(hash_sorted_pending_requests_.begin(),
+               hash_sorted_pending_requests_.end(),
+               std::back_inserter(output_list),
+               comparator);
+
+  return output_list;
 }
 
 bool RequestInfoSet::Erase(const RequestInfoPtr request_info) {
