+++ src/components/application_manager/src/state_controller_impl.cc
@@ -31,12 +31,13 @@
  */
 
 #include "application_manager/state_controller_impl.h"
+
 #include <tuple>
+
 #include "application_manager/rpc_service.h"
 #include "application_manager/usage_statistics.h"
-#include "utils/helpers.h"
-
 #include "connection_handler/connection_handler.h"
+#include "utils/helpers.h"
 
 namespace application_manager {
 
@@ -720,7 +721,8 @@ void StateControllerImpl::UpdateAppWindowsStreamingState(
   }
 }
 
-void StateControllerImpl::on_event(const event_engine::MobileEvent& event) {
+void StateControllerImpl::HandleOnEvent(
+    const event_engine::MobileEvent& event) {
   using namespace mobile_apis;
 
   SDL_LOG_AUTO_TRACE();
@@ -765,9 +767,9 @@ void StateControllerImpl::on_event(const event_engine::MobileEvent& event) {
     default:
       break;
   }
-}
+}  // namespace application_manager
 
-void StateControllerImpl::on_event(const event_engine::Event& event) {
+void StateControllerImpl::HandleOnEvent(const event_engine::Event& event) {
   using event_engine::Event;
   using smart_objects::SmartObject;
   using namespace hmi_apis;
